requirements.txt
streamlit
gitpython
sentence-transformers
faiss-cpu
numpy
tf-keras